Update/Delete user

+++ src/main/java/iit/cs445/models/users/UserService.java
@@ -20,7 +20,7 @@ public interface UserService {
         return null;
     }
 
-    default User findUserById(int id) {
+    default User findUserById(Long id) {
         return null;
     }
